+++ server/src/api/feed/constant.ts
@@ -1,0 +1,2 @@
+export const NEW_FEED = 'NEW_FEED_PUBSUB';
+export const NEW_ALARM = 'NEW_ALARM_PUBSUB';
