+++ server/src/api/feed/feedAlarm.resolvers.ts
@@ -1,0 +1,9 @@
+import queryResolvers from './feedAlarm.query';
+import mutationResolvers from './feedAlarm.mutation';
+import SubscriptionResolver from './feedAlarm.subscription';
+
+export default {
+  Query: queryResolvers,
+  Mutation: mutationResolvers,
+  Subscription: SubscriptionResolver
+};
